pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tank-lib"
version = "1.0.0"
description = "Core Tank functionality - controllers, models, and utilities"
authors = [
    {name = "Tank Team"}
]
requires-python = ">=3.12"
dependencies = [
    "boto3==1.35.38",
    "botocore==1.35.38",
    "Flask==3.1.0",
    "Flask-Cognito==1.21",
    "Flask-Caching==2.1.0",
    "PyJWT==2.10.1",
    "Requests==2.32.3",
    "validate_email==1.3",
    "cryptography==38.0.4",
    "openai==1.65.2",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
tank = ["app_chat/blueprints/*.json"]

